requirement.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
html-parser==0.2
idna==3.4
kombu==5.2.4
lxml==4.9.1
ply==3.11
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
